use UInt as id

+++ src/EGraphs/unionfind.jl
@@ -1,10 +1,10 @@
 struct UnionFind
-  parents::Vector{Int}
+  parents::Vector{UInt}
 end
 
-UnionFind() = UnionFind(Int[])
+UnionFind() = UnionFind(UInt[])
 
-function Base.push!(uf::UnionFind)
+function Base.push!(uf::UnionFind)::UInt
   l = length(uf.parents) + 1
   push!(uf.parents, l)
   l
@@ -12,12 +12,12 @@ end
 
 Base.length(uf::UnionFind) = length(uf.parents)
 
-function Base.union!(uf::UnionFind, i::Int, j::Int)
+function Base.union!(uf::UnionFind, i::UInt, j::UInt)
   uf.parents[j] = i
   i
 end
 
-function find(uf::UnionFind, i::Int)
+function find(uf::UnionFind, i::UInt)
   while i != uf.parents[i]
     i = uf.parents[i]
   end
